Refactor: trySignUp 함수 authVerifyUser로 이동
